Third Problem ...
This one was also pretty simple. I saw one solution that although cool, and efficient was so unreadable it made no sense. It took me a while to figure out what the hell was happening and in the end, it felt kinda wrong in the sense that I didn't like how it was just used the idea of returning boolean "numbers" where 1 or greater being true, and 0 or lower being false. Is it correct, and efficient absolutely! But it feels kinda hacky to me. Either way the logic was straightforward find the maximum in the List, and if there exists an element inside the List, such that the sum of that element + extraCandies is still less than the max, then append(False), otherwise append(True) into the Array. This main issue here was syntax plain and simple!